refs #46736 ColissimoRule
